initial comment
A number is given ,user has to guess the number and he has limited no. of guesses. After attending a guess it will show "no. of guesses left". If he guessed right then it will show succeed and no. of guesses he took to finish and if can not guess in the limited no. of guesses then it will show game over!